Add metadata to Wrapper
This commit adds TypeMeta and ObjectMeta to the Wrapper datatype.
It does not remove ObjectMeta from Sensu resources; rather, it
copies ObjectMeta into Wrappers when they are displayed, and hides
ObjectMeta in the resource.

When users specify ObjectMeta in Wrappers, the ObjectMeta will be
stored in the wrapped resource.

This commit also changes most of the resource protos by using the
gogoproto "face" extension. This was done so that GetObjectMeta
methods would be generated for resources embedding ObjectMeta.

The wrapper can now resolve types in multiple API groups, but
right now, only "core/v2" is supported. As a legacy fallback
mechanism, an empty API version means "core/v2" as well.

Some files have been updated to use the new v2 package naming
instead of the alias in the types package.

The Resource interface in the types package has been removed in
favour of an alias to v2.Resource.

Signed-off-by: Eric Chlebek <eric@sensu.io>